# internal/interface/http/handler/openapi.yaml
openapi: "3.0.3"
info:
  title: Stock Service API
  version: "1.0.0"
  description: API for managing inventory, stores, promotions, warehouses, and related resources.
paths:
  /api/v1/inventory:
    post:
      summary: Create a new inventory listing
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [store_id, warehouse_id, product_id, base_price]
              properties:
                store_id:     { type: integer, format: int64 }
                warehouse_id: { type: integer, format: int64 }
                product_id:   { type: integer, format: int32 }
                base_price:   { type: number, format: double }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      summary: List inventory with optional filters
      tags: [Inventory]
      parameters:
        - name: store_id
          in: query
          schema: { type: integer, format: int64 }
        - name: product_id
          in: query
          schema: { type: integer, format: int32 }
        - name: vendor_sale_status
          in: query
          schema: { type: string }
        - name: system_sale_status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
  /api/v1/inventory/search:
    get:
      summary: Search inventory by product name
      tags: [Inventory]
      parameters:
        - name: query
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/inventory/{id}:
    get:
      summary: Get inventory by ID
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete an inventory item
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/inventory/{id}/promotion:
    post:
      summary: Apply a promotion to an inventory item
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [promotion_id, final_price, start_at, end_at]
              properties:
                promotion_id: { type: integer, format: int64 }
                final_price:  { type: number, format: double }
                start_at:     { type: string, format: date-time }
                end_at:       { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      summary: Remove a promotion from an inventory item
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/inventory/{id}/inventory:
    put:
      summary: Update inventory stock quantities
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instant_qty, scheduled_qty, min_order_qty]
              properties:
                instant_qty:   { type: integer }
                scheduled_qty:
                  type: object
                  additionalProperties: { type: integer }
                min_order_qty: { type: integer }
                max_order_qty: { type: integer, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/inventory/{id}/vendor/suspend:
    post:
      summary: Suspend vendor sale
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/inventory/{id}/vendor/close:
    post:
      summary: Close vendor sale
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/inventory/{id}/system/suspend:
    post:
      summary: Suspend system sale
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/inventory/{id}/system/close:
    post:
      summary: Close system sale
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/inventory/{id}/reserve:
    post:
      summary: Reserve quantity from inventory
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/inventory/{id}/release:
    post:
      summary: Release reserved quantity back to inventory
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/inventory/{id}/low-stock:
    get:
      summary: Check if inventory is low on stock
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: threshold
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LowStockResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    InventoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryResponse"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    LowStockResponse:
      type: object
      properties:
        is_low:
          type: boolean
        current_qty:
          type: integer
    InventoryResponse:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        StoreID:
          type: integer
          format: int64
        WarehouseID:
          type: integer
          format: int64
        ProductID:
          type: integer
          format: int32
        SaleModel:
          type: string
        BasePrice:
          type: number
          format: double
        PromotionID:
          type: integer
          format: int64
          nullable: true
        FinalPrice:
          type: number
          format: double
          nullable: true
        StartAt:
          type: string
          format: date-time
          nullable: true
        EndAt:
          type: string
          format: date-time
          nullable: true
        PromotionStatus:
          type: string
        InstantQty:
          type: integer
        MinOrderQty:
          type: integer
        MaxOrderQty:
          type: integer
          nullable: true
        Condition:
          type: string
        VendorSaleStatus:
          type: string
        SystemSaleStatus:
          type: string
        CreatedAt:
          type: string
          format: date-time
...